requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-filter>=23.5
drf-spectacular>=0.27
django-cors-headers>=4.3
django-jazzmin>=3.0
PyJWT>=2.8
psycopg2-binary>=2.9
channels>=4.0,<5.0
channels-redis>=4.2
daphne>=4.0
